+++ flask_unchained/bundles/sqlalchemy/meta_options.py
@@ -1,13 +1,12 @@
 from flask_unchained import unchained
 from py_meta_utils import McsArgs, MetaOption
-from sqlalchemy.orm import RelationshipProperty
 from sqlalchemy_unchained import ModelMetaOptionsFactory as BaseModelMetaOptionsFactory
 from typing import *
 
 
 class ModelMetaOption(MetaOption):
     """
-    The model class this model resource is for.
+    The model class for a class.
     """
     def __init__(self):
         super().__init__('model', default=None, inherit=True)
@@ -27,42 +26,6 @@ class ModelMetaOption(MetaOption):
             raise TypeError(f'{mcs_args.name} is missing the model Meta attribute')
 
 
-class RelationshipsMetaOption(MetaOption):
-    def __init__(self):
-        super().__init__('relationships', inherit=True)
-
-    def get_value(self, meta, base_model_meta, mcs_args: McsArgs):
-        """overridden to merge with inherited value"""
-        if mcs_args.Meta.abstract:
-            return None
-        value = getattr(base_model_meta, self.name, {}) or {}
-        value.update(getattr(meta, self.name, {}))
-        return value
-
-    def contribute_to_class(self, mcs_args: McsArgs, relationships):
-        if mcs_args.Meta.abstract:
-            return
-
-        discovered_relationships = {}
-
-        def discover_relationships(d):
-            for k, v in d.items():
-                if isinstance(v, RelationshipProperty):
-                    discovered_relationships[v.argument] = k
-                    if v.backref and mcs_args.Meta.lazy_mapped:
-                        raise NotImplementedError(
-                            f'Discovered a lazy-mapped backref `{k}` on '
-                            f'`{mcs_args.qualname}`. Currently this '
-                            'is unsupported; please use `db.relationship` with '
-                            'the `back_populates` kwarg on both sides instead.')
-
-        for base in mcs_args.bases:
-            discover_relationships(vars(base))
-        discover_relationships(mcs_args.clsdict)
-
-        relationships.update(discovered_relationships)
-
-
 class MaterializedViewForMetaOption(MetaOption):
     def __init__(self):
         super().__init__(name='mv_for', default=None, inherit=True)
@@ -73,5 +36,4 @@ class MaterializedViewForMetaOption(MetaOption):
 
 class ModelMetaOptionsFactory(BaseModelMetaOptionsFactory):
     def _get_meta_options(self) -> List[MetaOption]:
-        return super()._get_meta_options() + [RelationshipsMetaOption(),
-                                              MaterializedViewForMetaOption()]
+        return super()._get_meta_options() + [MaterializedViewForMetaOption()]