+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fix auto-generated imports in db migrations requiring user app code
 - add compatibility with Flask-Session 0.4.0
 - fix compatibility with Flask-WTF auto-population of form data on submit
+- move some more database code into sqlalchemy-unchained
 
 ## v0.9.0 (2021/06/07)